n-ipv4ll: increase entropy when selecting enumeration
To reduce the likelihood of two instances of n-ipv4ll selecting the precise
same enumeration (and hence not terminating) switch to using a proper prng.
This increases the entropy from 16 to 48 bits, at the cost of potentially
re-trying the same conflicting IP address before having exhausted the pool.
This however is not a real concern as detecting a conflict is fast (only
verifying there is no conflict is slow).

This also changes n_ipv4ll_get_ip() to fail in case the state machine
was never started.

Signed-off-by: Tom Gundersen <teg@jklm.no>